Apply the changes in foundry-keystore to CLI
foundry-primitives@0.2.1 and foundry-keystore@0.1.0 were published.
We removed unnecessary the Asset type from them and renamed the Platform type to 'keystore'.
Therefore, there is no keytype in the renewed keystore.
In addition, they use the Ed25519 signature, which uses 64 bytes private key.

To apply the changes in foundry-keystore and primitives to foundry-cli,
we removed the concept of accountType, renamed it to 'keystore'
and changed 32 bytes key private key to 64 bytes.